# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Fact:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      restaurant_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@restaurant-booking.com
    name: API Support
  description: API для системы бронирования ресторанов
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Booking API
  version: "1.0"
paths:
...
  /facts/random:
    get:
      consumes:
      - application/json
      description: Get a collection of random facts about restaurants
      parameters:
      - default: 3
        description: Number of facts to return
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Fact'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get random facts
      tags:
      - facts
...
schemes:
- http
- https
swagger: "2.0"